Simplify archetype type storage to use generated type more directly, refine type generation to reverse order by size to avoid alignment issues on type retrieval, update tests to check for alignment issues

+++ examples/pong/main.zig
@@ -10,7 +10,7 @@ pub const c = @cImport({
 });
 
 // Component type declarations
-const Id = struct { id: u32 };
+const Id = struct { id: u8 };
 const Movespeed = struct { speed: u32 };
 const Position = struct { x: i32, y: i32 };
 const Size = struct { w: i32, h: i32 };
